REMOVE: EntityInfo
