+++ src/lib/kotlin/slatekit-entities/src/main/kotlin/slatekit/entities/core/EntityUtils.kt
@@ -1,0 +1,16 @@
+package slatekit.entities.slatekit.entities.core
+
+import java.util.*
+
+object EntityUtils {
+
+    fun <TId> isCreated(id: TId): Boolean {
+        return when (id) {
+            is Int -> id > 0
+            is Long -> id > 0L
+            is String -> !id.isEmpty()
+            is UUID -> !id.toString().trim().isEmpty()
+            else -> false
+        }
+    }
+}
